`test-runner-core`: add reconnection mechanism for Browserstack on `testsStartTimeout`
iOS tests sometimes fail to start due to some "server disconnect" issue on BS side.
This seems more prominent on certain devices (e.g. iPhone 16 with iOS 18).
Root cause is unclear, and refreshing the page (a.k.a. simply restarting the test session) is not sufficient: reconnecting the launcher is needed.